Update: Modularise all Slicey-related components
