+++ src/components/alexandria/Slicey/Donut/styles.scss
@@ -1,4 +1,4 @@
-@import '../../color';
+@import 'styles/color';
 
 .donut-component {
   fill: $color-white;
