+++ src/components/alexandria/Slicey/Marker/styles.scss
@@ -1,4 +1,4 @@
-@import '../../color';
+@import 'styles/color';
 
 .marker-component {
   stroke: $color-red;
