+++ src/components/alexandria/Slicey/styles.scss
@@ -1,4 +1,4 @@
-@import '../color';
+@import 'styles/color';
 
 .slicey {
   &-component {
